collectors/trades_collector/requirements.txt
APScheduler
binance-connector
redis